+++ lib/active_remote/integration.rb
@@ -5,8 +5,10 @@ module ActiveRemote
     included do
       ##
       # :singleton-method:
-      # Indicates the format used to generate the timestamp in the cache key, if
-      # versioning is off. Accepts any of the symbols in <tt>Time::DATE_FORMATS</tt>.
+      # Indicates the format used to generate the timestamp #cache_key appends
+      # when ActiveRemote.config.default_cache_key_updated_at? is set. Accepts
+      # any of the symbols in <tt>Time::DATE_FORMATS</tt>. #cache_version does
+      # not consult this and always uses +:usec+.
       #
       # This is +:usec+, by default.
       class_attribute :cache_timestamp_format, instance_writer: false, default: :usec
@@ -16,25 +18,25 @@ module ActiveRemote
       # Indicates whether to use a stable #cache_key method that is accompanied
       # by a changing version in the #cache_version method.
       #
-      # This is +false+, by default until Rails 6.0.
+      # This is +false+, by default.
       class_attribute :cache_versioning, instance_writer: false, default: false
     end
 
     # Returns a +String+, which Action Pack uses for constructing a URL to this
-    # object. The default implementation returns this record's id as a +String+,
-    # or +nil+ if this record's unsaved.
+    # object. The default implementation returns this record's primary key as a
+    # +String+, or +nil+ if that key has no value.
     #
     # For example, suppose that you have a User model, and that you have a
     # <tt>resources :users</tt> route. Normally, +user_path+ will
-    # construct a path with the user object's 'id' in it:
+    # construct a path with the user object's primary key in it:
     #
     #   user = User.find_by(name: 'Phusion')
-    #   user_path(user)  # => "/users/1"
+    #   user_path(user)  # => "/users/ABC-123"
     #
     # You can override +to_param+ in your model to make +user_path+ construct
-    # a path using the user's name instead of the user's id:
+    # a path using the user's name instead:
     #
-    #   class User < ActiveRecord::Base
+    #   class User < ActiveRemote::Base
     #     def to_param  # overridden
     #       name
     #     end
@@ -48,16 +50,18 @@ module ActiveRemote
       key&.to_s
     end
 
-    # Returns a stable cache key that can be used to identify this record.
+    # Returns a stable cache key that can be used to identify this record. The
+    # key is built from the record's primary key, not an id.
     #
-    #   Product.new.cache_key     # => "products/new"
-    #   Product.find(5).cache_key # => "products/5"
+    #   Product.new.cache_key # => "products/new"
+    #   product.cache_key     # => "products/ABC-123"
     #
-    # If ActiveRecord::Base.cache_versioning is turned off, as it was in Rails 5.1 and earlier,
-    # the cache key will also include a version.
+    # When ActiveRemote.config.default_cache_key_updated_at is set and the
+    # record carries an +updated_at+, the key also includes that timestamp,
+    # formatted with .cache_timestamp_format.
     #
-    #   Product.cache_versioning = false
-    #   Person.find(5).cache_key  # => "people/5-20071224150000" (updated_at available)
+    #   ActiveRemote.config.default_cache_key_updated_at = true
+    #   product.cache_key # => "products/ABC-123-20071224150000000000"
     #
     def cache_key
       if new_record?
@@ -84,8 +88,8 @@ module ActiveRemote
     # a recyclable caching scheme. By default, the #updated_at column is used for the
     # cache_version, but this method can be overwritten to return something else.
     #
-    # Note, this method will return nil if ActiveRecord::Base.cache_versioning is set to
-    # +false+ (which it is by default until Rails 6.0).
+    # Note, this method will return nil unless .cache_versioning is set to
+    # +true+ (it defaults to +false+).
     def cache_version
       if cache_versioning && (timestamp = try(:updated_at))
         timestamp.utc.to_fs(:usec)
@@ -97,26 +101,26 @@ module ActiveRemote
       # using +method_name+, which can be any attribute or method that
       # responds to +to_s+.
       #
-      #   class User < ActiveRecord::Base
+      #   class User < ActiveRemote::Base
       #     to_param :name
       #   end
       #
       #   user = User.find_by(name: 'Fancy Pants')
-      #   user.id         # => 123
+      #   user.guid       # => "123"
       #   user_path(user) # => "/users/123-fancy-pants"
       #
       # Values longer than 20 characters will be truncated. The value
       # is truncated word by word.
       #
       #   user = User.find_by(name: 'David Heinemeier Hansson')
-      #   user.id         # => 125
+      #   user.guid       # => "125"
       #   user_path(user) # => "/users/125-david-heinemeier"
       #
-      # Because the generated param begins with the record's +id+, it is
-      # suitable for passing to +find+. In a controller, for example:
+      # The generated param begins with the primary key but is not itself a
+      # valid argument to +find+, which takes a hash of search args.
       #
-      #   params[:id]               # => "123-fancy-pants"
-      #   User.find(params[:id]).id # => 123
+      #   params[:id] # => "123-fancy-pants"
+      #   User.find(guid: params[:id].split("-").first)
       def to_param(method_name = nil)
         if method_name.nil?
           super()