+++ spec/support/helpers.rb
@@ -13,7 +13,6 @@ def reset_dsl_variables(klass)
   reset_app_name(klass)
   reset_auto_paging_size(klass)
   reset_namespace(klass)
-  reset_publishable_attributes(klass)
   reset_service_class(klass)
   reset_service_name(klass)
 end
@@ -28,10 +27,6 @@ end
 
 def reset_namespace(klass)
   klass.send(:instance_variable_set, :@namespace, nil)
-end
-
-def reset_publishable_attributes(klass)
-  klass.send(:instance_variable_set, :@publishable_attributes, nil)
 end
 
 def reset_service_class(klass)